+++ src/util/coupons/calculateCoupons.ts
@@ -5,33 +5,8 @@ import { calculateTotalPrice } from "../cart/calculateTotalPrice";
 import { calculateBuyXGetYCoupon } from "./BuyXGetYCoupon/calculate";
 import { calculateFixedDiscountCoupon } from "./FixedDiscountCoupon/calculate";
 import { calculateFreeShippingCoupon } from "./FreeShippingCoupon/calculate";
+import { getCouponCombos } from "./getCouponCombos";
 import { calculatePercentageDiscountCoupon } from "./PercentageDiscountCoupon/calculate";
-
-const getCouponCombos = (coupons: Coupon[]) => {
-  const result: Coupon[][] = [];
-
-  function getPermutations(arr: Coupon[], currentPermutation: Coupon[]) {
-    if (arr.length === 0) {
-      result.push(currentPermutation);
-      return;
-    }
-
-    for (let i = 0; i < arr.length; i += 1) {
-      {
-        const nextElement = arr[i];
-        const remainingElements = [...arr.slice(0, i), ...arr.slice(i + 1)];
-        getPermutations(remainingElements, [
-          ...currentPermutation,
-          nextElement,
-        ]);
-      }
-    }
-  }
-
-  getPermutations(coupons, []);
-
-  return result;
-};
 
 const getDiscountAmount = ({
   coupon,
